pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["kian <kian.sa2020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
django-cors-headers = "^3.14.0"
markdown = "^3.4.3"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
